+++ dev/archetypes/maven-java/src/main/resources/archetype-resources/__service1Name__-api/src/main/java/__packageInPathFormat__/__service1Name__/api/GreetingMessage.java
@@ -3,7 +3,6 @@
  */
 package ${package}.${service1Name}.api;
 
-import javax.annotation.Nullable;
 import lombok.Value;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
