+++ dev/archetypes/maven-java/src/main/resources/archetype-resources/__service1Name__-impl/pom.xml
@@ -40,10 +40,10 @@
             <artifactId>lagom-javadsl-testkit_2.11</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
+<!--        <dependency>
             <groupId>com.typesafe.conductr</groupId>
             <artifactId>${conductr.lib.name}</artifactId>
-        </dependency>
+        </dependency> -->
        <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
