+++ dev/archetypes/maven-java/src/main/resources/archetype-resources/__service1Name__-impl/src/main/java/__packageInPathFormat__/__service1Name__/impl/__service1ClassName__Command.java
@@ -5,7 +5,6 @@ package ${package}.${service1Name}.impl;
 
 import java.util.Optional;
 
-import javax.annotation.Nullable;
 import lombok.Value;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
