+++ dev/archetypes/maven-java/src/main/resources/archetype-resources/__service1Name__-impl/src/main/java/__packageInPathFormat__/__service1Name__/impl/__service1ClassName__Event.java
@@ -3,7 +3,6 @@
  */
 package ${package}.${service1Name}.impl;
 
-import javax.annotation.Nullable;
 import lombok.Value;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
