+++ dev/archetypes/maven-java/src/main/resources/archetype-resources/__service2Name__-impl/pom.xml
@@ -41,10 +41,11 @@
             <artifactId>lagom-javadsl-testkit_2.11</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
+<!--        <dependency>
             <groupId>com.typesafe.conductr</groupId>
             <artifactId>${conductr.lib.name}</artifactId>
         </dependency>
+        -->
     </dependencies>
 
     <build>
