+++ dev/sbt-plugin/src/sbt-test/sbt-plugin/distribution/playProj/app/views/index.scala.html
@@ -1,7 +1,4 @@
 @(message: String)
 
 @main("Welcome to Play") {
-
-    @play20.welcome(message)
-
 }
